rassemblement equipementStore et roomStore
